Moved block variant card to a separate file
